@@ -5,7 +5,7 @@ author: Charts Team <charts_flutter@google.com>
 homepage: https://github.com/google/charts
 
 environment:
-  sdk: '>=1.23.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   # Pointing this to a local path allows for pointing to the latest code
