@@ -44,34 +44,34 @@ abstract class BaseChart<D> extends StatefulWidget {
 
   /// Used to configure the margin sizes around the drawArea that the axis and
   /// other things render into.
-  final LayoutConfig layoutConfig;
+  final LayoutConfig? layoutConfig;
 
   // Default renderer used to draw series data on the chart.
-  final common.SeriesRendererConfig<D> defaultRenderer;
+  final common.SeriesRendererConfig<D>? defaultRenderer;
 
   /// Include the default interactions or not.
   final bool defaultInteractions;
 
-  final List<ChartBehavior> behaviors;
+  final List<ChartBehavior>? behaviors;
 
-  final List<SelectionModelConfig<D>> selectionModels;
+  final List<SelectionModelConfig<D?>>? selectionModels;
 
   // List of custom series renderers used to draw series data on the chart.
   //
   // Series assigned a rendererIdKey will be drawn with the matching renderer in
   // this list. Series without a rendererIdKey will be drawn by the default
   // renderer.
-  final List<common.SeriesRendererConfig<D>> customSeriesRenderers;
+  final List<common.SeriesRendererConfig<D>>? customSeriesRenderers;
 
   /// The spec to use if RTL is enabled.
-  final common.RTLSpec rtlSpec;
+  final common.RTLSpec? rtlSpec;
 
   /// Optional state that overrides internally kept state, such as selection.
-  final UserManagedState<D> userManagedState;
+  final UserManagedState<D>? userManagedState;
 
   BaseChart(this.seriesList,
-      {bool animate,
-      Duration animationDuration,
+      {bool? animate,
+      Duration? animationDuration,
       this.defaultRenderer,
       this.customSeriesRenderers,
       this.behaviors,
@@ -88,16 +88,16 @@ abstract class BaseChart<D> extends StatefulWidget {
   BaseChartState<D> createState() => new BaseChartState<D>();
 
   /// Creates and returns a [common.BaseChart].
-  common.BaseChart<D> createCommonChart(BaseChartState<D> chartState);
+  common.BaseChart<D> createCommonChart(BaseChartState<D>? chartState);
 
   /// Updates the [common.BaseChart].
-  void updateCommonChart(common.BaseChart chart, BaseChart<D> oldWidget,
+  void updateCommonChart(common.BaseChart chart, BaseChart<D>? oldWidget,
       BaseChartState<D> chartState) {
     common.Performance.time('chartsUpdateRenderers');
     // Set default renderer if one was provided.
     if (defaultRenderer != null &&
         defaultRenderer != oldWidget?.defaultRenderer) {
-      chart.defaultRenderer = defaultRenderer.build();
+      chart.defaultRenderer = defaultRenderer!.build();
       chartState.markChartDirty();
     }
 
@@ -105,12 +105,12 @@ abstract class BaseChart<D> extends StatefulWidget {
     if (customSeriesRenderers != null) {
       // TODO: This logic does not remove old renderers and
       // shouldn't require the series configs to remain in the same order.
-      for (var i = 0; i < customSeriesRenderers.length; i++) {
+      for (var i = 0; i < customSeriesRenderers!.length; i++) {
         if (oldWidget == null ||
             (oldWidget.customSeriesRenderers != null &&
-                i > oldWidget.customSeriesRenderers.length) ||
-            customSeriesRenderers[i] != oldWidget.customSeriesRenderers[i]) {
-          chart.addSeriesRenderer(customSeriesRenderers[i].build());
+                i > oldWidget.customSeriesRenderers!.length) ||
+            customSeriesRenderers![i] != oldWidget.customSeriesRenderers![i]) {
+          chart.addSeriesRenderer(customSeriesRenderers![i].build());
           chartState.markChartDirty();
         }
       }
@@ -121,14 +121,14 @@ abstract class BaseChart<D> extends StatefulWidget {
     _updateBehaviors(chart, chartState);
     common.Performance.timeEnd('chartsUpdateBehaviors');
 
-    _updateSelectionModel(chart, chartState);
+    _updateSelectionModel(chart as BaseChart<D?>, chartState);
 
     chart.transition = animate ? animationDuration : Duration.zero;
   }
 
   void _updateBehaviors(common.BaseChart chart, BaseChartState chartState) {
     final behaviorList = behaviors != null
-        ? new List<ChartBehavior>.from(behaviors)
+        ? new List<ChartBehavior>.from(behaviors!)
         : <ChartBehavior>[];
 
     // Insert automatic behaviors to the front of the behavior list.
@@ -155,7 +155,7 @@ abstract class BaseChart<D> extends StatefulWidget {
         final role = addedBehavior.role;
         chartState.addedBehaviorWidgets.remove(addedBehavior);
         chartState.addedCommonBehaviorsByRole.remove(role);
-        chart.removeBehavior(chartState.addedCommonBehaviorsByRole[role]);
+        chart.removeBehavior(chartState.addedCommonBehaviorsByRole[role]!);
         chartState.markChartDirty();
       }
     }
@@ -163,7 +163,7 @@ abstract class BaseChart<D> extends StatefulWidget {
     // Add any remaining/new behaviors.
     behaviorList.forEach((ChartBehavior behaviorWidget) {
       final commonBehavior = chart
-          .createBehavior(<D>() => behaviorWidget.createCommonBehavior<D>());
+          .createBehavior(<D>() => behaviorWidget.createCommonBehavior<D>() as ChartBehavior<D>);
 
       // Assign the chart state to any behavior that needs it.
       if (commonBehavior is ChartStateBehavior) {
@@ -189,24 +189,24 @@ abstract class BaseChart<D> extends StatefulWidget {
 
   bool _notACustomBehavior(ChartBehavior behavior) {
     return this.behaviors == null ||
-        !this.behaviors.any(
+        !this.behaviors!.any(
             (ChartBehavior userBehavior) => userBehavior.role == behavior.role);
   }
 
   void _updateSelectionModel(
-      common.BaseChart<D> chart, BaseChartState<D> chartState) {
+      common.BaseChart<D?> chart, BaseChartState<D?> chartState) {
     final prevTypes = new List<common.SelectionModelType>.from(
         chartState.addedSelectionChangedListenersByType.keys);
 
     // Update any listeners for each type.
-    selectionModels?.forEach((SelectionModelConfig<D> model) {
+    selectionModels?.forEach((SelectionModelConfig<D?> model) {
       final selectionModel = chart.getSelectionModel(model.type);
 
       final prevChangedListener =
           chartState.addedSelectionChangedListenersByType[model.type];
       if (!identical(model.changedListener, prevChangedListener)) {
-        selectionModel.removeSelectionChangedListener(prevChangedListener);
-        selectionModel.addSelectionChangedListener(model.changedListener);
+        selectionModel.removeSelectionChangedListener(prevChangedListener!);
+        selectionModel.addSelectionChangedListener(model.changedListener!);
         chartState.addedSelectionChangedListenersByType[model.type] =
             model.changedListener;
       }
@@ -214,8 +214,8 @@ abstract class BaseChart<D> extends StatefulWidget {
       final prevUpdatedListener =
           chartState.addedSelectionUpdatedListenersByType[model.type];
       if (!identical(model.updatedListener, prevUpdatedListener)) {
-        selectionModel.removeSelectionUpdatedListener(prevUpdatedListener);
-        selectionModel.addSelectionUpdatedListener(model.updatedListener);
+        selectionModel.removeSelectionUpdatedListener(prevUpdatedListener!);
+        selectionModel.addSelectionUpdatedListener(model.updatedListener!);
         chartState.addedSelectionUpdatedListenersByType[model.type] =
             model.updatedListener;
       }
@@ -227,9 +227,9 @@ abstract class BaseChart<D> extends StatefulWidget {
     prevTypes.forEach((common.SelectionModelType type) {
       chart.getSelectionModel(type)
         ..removeSelectionChangedListener(
-            chartState.addedSelectionChangedListenersByType[type])
+            chartState.addedSelectionChangedListenersByType[type]!)
         ..removeSelectionUpdatedListener(
-            chartState.addedSelectionUpdatedListenersByType[type]);
+            chartState.addedSelectionUpdatedListenersByType[type]!);
     });
   }
 
@@ -242,7 +242,7 @@ abstract class BaseChart<D> extends StatefulWidget {
   Set<GestureType> getDesiredGestures(BaseChartState chartState) {
     final types = new Set<GestureType>();
     behaviors?.forEach((ChartBehavior behavior) {
-      types.addAll(behavior.desiredGestures);
+      types.addAll(behavior.desiredGestures!);
     });
 
     if (defaultInteractions && chartState.autoBehaviorWidgets.isEmpty) {
@@ -250,7 +250,7 @@ abstract class BaseChart<D> extends StatefulWidget {
     }
 
     chartState.autoBehaviorWidgets.forEach((ChartBehavior behavior) {
-      types.addAll(behavior.desiredGestures);
+      types.addAll(behavior.desiredGestures!);
     });
     return types;
   }
@@ -264,10 +264,10 @@ class LayoutConfig {
   final common.MarginSpec bottomMarginSpec;
 
   LayoutConfig({
-    @required this.leftMarginSpec,
-    @required this.topMarginSpec,
-    @required this.rightMarginSpec,
-    @required this.bottomMarginSpec,
+    required this.leftMarginSpec,
+    required this.topMarginSpec,
+    required this.rightMarginSpec,
+    required this.bottomMarginSpec,
   });
 
   common.LayoutConfig get commonLayoutConfig => new common.LayoutConfig(
