@@ -31,25 +31,25 @@ import '../chart_behavior.dart' show ChartBehavior, GestureType;
 class ChartTitle extends ChartBehavior<common.ChartTitle> {
   final desiredGestures = new Set<GestureType>();
 
-  final common.BehaviorPosition behaviorPosition;
+  final common.BehaviorPosition? behaviorPosition;
 
   /// Minimum size of the legend component. Optional.
   ///
   /// If the legend is positioned in the top or bottom margin, then this
   /// configures the legend's height. If positioned in the start or end
   /// position, this configures the legend's width.
-  final int layoutMinSize;
+  final int? layoutMinSize;
 
   /// Preferred size of the legend component. Defaults to 0.
   ///
   /// If the legend is positioned in the top or bottom margin, then this
   /// configures the legend's height. If positioned in the start or end
   /// position, this configures the legend's width.
-  final int layoutPreferredSize;
+  final int? layoutPreferredSize;
 
   /// Strategy for handling title text that is too large to fit. Defaults to
   /// truncating the text with ellipses.
-  final common.MaxWidthStrategy maxWidthStrategy;
+  final common.MaxWidthStrategy? maxWidthStrategy;
 
   /// Primary text for the title.
   final String title;
@@ -58,28 +58,28 @@ class ChartTitle extends ChartBehavior<common.ChartTitle> {
   ///
   /// This defaults to horizontal for a title in the top or bottom
   /// [behaviorPosition], or vertical for start or end [behaviorPosition].
-  final common.ChartTitleDirection titleDirection;
+  final common.ChartTitleDirection? titleDirection;
 
   /// Justification of the title text if it is positioned outside of the draw
   /// area.
-  final common.OutsideJustification titleOutsideJustification;
+  final common.OutsideJustification? titleOutsideJustification;
 
   /// Space between the title and sub-title text, if defined.
   ///
   /// This padding is not used if no sub-title is provided.
-  final int titlePadding;
+  final int? titlePadding;
 
   /// Style of the [title] text.
-  final common.TextStyleSpec titleStyleSpec;
+  final common.TextStyleSpec? titleStyleSpec;
 
   /// Secondary text for the sub-title.
   ///
   /// [subTitle] is rendered on a second line below the [title], and may be
   /// styled differently.
-  final String subTitle;
+  final String? subTitle;
 
   /// Style of the [subTitle] text.
-  final common.TextStyleSpec subTitleStyleSpec;
+  final common.TextStyleSpec? subTitleStyleSpec;
 
   /// Space between the "inside" of the chart, and the title behavior itself.
   ///
@@ -91,7 +91,7 @@ class ChartTitle extends ChartBehavior<common.ChartTitle> {
   /// If a sub-title is defined, this is the space between the sub-title text
   /// and the inside of the chart. Otherwise, it is the space between the title
   /// text and the inside of chart.
-  final int innerPadding;
+  final int? innerPadding;
 
   /// Space between the "outside" of the chart, and the title behavior itself.
   ///
@@ -99,7 +99,7 @@ class ChartTitle extends ChartBehavior<common.ChartTitle> {
   /// direction of the draw area. For a top positioned title, this is applied
   /// to the top, left, and right edges. [innerPadding] is applied to the
   /// bottom edge.
-  final int outerPadding;
+  final int? outerPadding;
 
   /// Constructs a [ChartTitle].
   ///
@@ -143,18 +143,18 @@ class ChartTitle extends ChartBehavior<common.ChartTitle> {
   @override
   common.ChartTitle<D> createCommonBehavior<D>() =>
       new common.ChartTitle<D>(title,
-          behaviorPosition: behaviorPosition,
-          innerPadding: innerPadding,
-          layoutMinSize: layoutMinSize,
-          layoutPreferredSize: layoutPreferredSize,
-          outerPadding: outerPadding,
-          maxWidthStrategy: maxWidthStrategy,
-          titleDirection: titleDirection,
-          titleOutsideJustification: titleOutsideJustification,
-          titlePadding: titlePadding,
-          titleStyleSpec: titleStyleSpec,
-          subTitle: subTitle,
-          subTitleStyleSpec: subTitleStyleSpec);
+          behaviorPosition: behaviorPosition!,
+          innerPadding: innerPadding!,
+          layoutMinSize: layoutMinSize!,
+          layoutPreferredSize: layoutPreferredSize!,
+          outerPadding: outerPadding!,
+          maxWidthStrategy: maxWidthStrategy!,
+          titleDirection: titleDirection!,
+          titleOutsideJustification: titleOutsideJustification!,
+          titlePadding: titlePadding!,
+          titleStyleSpec: titleStyleSpec!,
+          subTitle: subTitle!,
+          subTitleStyleSpec: subTitleStyleSpec!);
 
   @override
   void updateCommonBehavior(common.ChartTitle commonBehavior) {}
