@@ -48,17 +48,17 @@ class DatumLegend extends ChartBehavior<common.DatumLegend> {
 
   final desiredGestures = new Set<GestureType>();
 
-  final common.SelectionModelType selectionModelType;
+  final common.SelectionModelType? selectionModelType;
 
   /// Builder for creating custom legend content.
-  final LegendContentBuilder contentBuilder;
+  final LegendContentBuilder? contentBuilder;
 
   /// Position of the legend relative to the chart.
-  final common.BehaviorPosition position;
+  final common.BehaviorPosition? position;
 
   /// Justification of the legend relative to the chart
-  final common.OutsideJustification outsideJustification;
-  final common.InsideJustification insideJustification;
+  final common.OutsideJustification? outsideJustification;
+  final common.InsideJustification? insideJustification;
 
   /// Whether or not the legend should show measures.
   ///
@@ -69,23 +69,23 @@ class DatumLegend extends ChartBehavior<common.DatumLegend> {
   ///
   /// This flag is used by the [contentBuilder], so a custom content builder
   /// has to choose if it wants to use this flag.
-  final bool showMeasures;
+  final bool? showMeasures;
 
   /// Option to show measures when selection is null.
   ///
   /// By default this is set to none, so no measures are shown when there is
   /// no selection.
-  final common.LegendDefaultMeasure legendDefaultMeasure;
+  final common.LegendDefaultMeasure? legendDefaultMeasure;
 
   /// Formatter for measure value(s) if the measures are shown on the legend.
-  final common.MeasureFormatter measureFormatter;
+  final common.MeasureFormatter? measureFormatter;
 
   /// Formatter for secondary measure value(s) if the measures are shown on the
   /// legend and the series uses the secondary axis.
-  final common.MeasureFormatter secondaryMeasureFormatter;
+  final common.MeasureFormatter? secondaryMeasureFormatter;
 
   /// Styles for legend entry label text.
-  final common.TextStyleSpec entryTextStyle;
+  final common.TextStyleSpec? entryTextStyle;
 
   static const defaultCellPadding = const EdgeInsets.all(8.0);
 
@@ -127,18 +127,18 @@ class DatumLegend extends ChartBehavior<common.DatumLegend> {
   /// [secondaryMeasureFormatter] formats measures if measures are shown for the
   /// series that uses secondary measure axis.
   factory DatumLegend({
-    common.BehaviorPosition position,
-    common.OutsideJustification outsideJustification,
-    common.InsideJustification insideJustification,
-    bool horizontalFirst,
-    int desiredMaxRows,
-    int desiredMaxColumns,
-    EdgeInsets cellPadding,
-    bool showMeasures,
-    common.LegendDefaultMeasure legendDefaultMeasure,
-    common.MeasureFormatter measureFormatter,
-    common.MeasureFormatter secondaryMeasureFormatter,
-    common.TextStyleSpec entryTextStyle,
+    common.BehaviorPosition? position,
+    common.OutsideJustification? outsideJustification,
+    common.InsideJustification? insideJustification,
+    bool? horizontalFirst,
+    int? desiredMaxRows,
+    int? desiredMaxColumns,
+    EdgeInsets? cellPadding,
+    bool? showMeasures,
+    common.LegendDefaultMeasure? legendDefaultMeasure,
+    common.MeasureFormatter? measureFormatter,
+    common.MeasureFormatter? secondaryMeasureFormatter,
+    common.TextStyleSpec? entryTextStyle,
   }) {
     // Set defaults if empty.
     position ??= defaultBehaviorPosition;
@@ -201,14 +201,14 @@ class DatumLegend extends ChartBehavior<common.DatumLegend> {
   /// series that uses secondary measure axis.
   factory DatumLegend.customLayout(
     LegendContentBuilder contentBuilder, {
-    common.BehaviorPosition position,
-    common.OutsideJustification outsideJustification,
-    common.InsideJustification insideJustification,
-    bool showMeasures,
-    common.LegendDefaultMeasure legendDefaultMeasure,
-    common.MeasureFormatter measureFormatter,
-    common.MeasureFormatter secondaryMeasureFormatter,
-    common.TextStyleSpec entryTextStyle,
+    common.BehaviorPosition? position,
+    common.OutsideJustification? outsideJustification,
+    common.InsideJustification? insideJustification,
+    bool? showMeasures,
+    common.LegendDefaultMeasure? legendDefaultMeasure,
+    common.MeasureFormatter? measureFormatter,
+    common.MeasureFormatter? secondaryMeasureFormatter,
+    common.TextStyleSpec? entryTextStyle,
   }) {
     // Set defaults if empty.
     position ??= defaultBehaviorPosition;
@@ -295,12 +295,12 @@ class _FlutterDatumLegend<D> extends common.DatumLegend<D>
 
   _FlutterDatumLegend(this.config)
       : super(
-          selectionModelType: config.selectionModelType,
-          measureFormatter: config.measureFormatter,
-          secondaryMeasureFormatter: config.secondaryMeasureFormatter,
-          legendDefaultMeasure: config.legendDefaultMeasure,
+          selectionModelType: config.selectionModelType!,
+          measureFormatter: config.measureFormatter!,
+          secondaryMeasureFormatter: config.secondaryMeasureFormatter!,
+          legendDefaultMeasure: config.legendDefaultMeasure!,
         ) {
-    super.entryTextStyle = config.entryTextStyle;
+    super.entryTextStyle = config.entryTextStyle!;
   }
 
   @override
@@ -309,15 +309,15 @@ class _FlutterDatumLegend<D> extends common.DatumLegend<D>
   }
 
   @override
-  common.BehaviorPosition get position => config.position;
+  common.BehaviorPosition? get position => config.position;
 
   @override
   common.OutsideJustification get outsideJustification =>
-      config.outsideJustification;
+      config.outsideJustification!;
 
   @override
   common.InsideJustification get insideJustification =>
-      config.insideJustification;
+      config.insideJustification!;
 
   @override
   Widget build(BuildContext context) {
@@ -326,11 +326,11 @@ class _FlutterDatumLegend<D> extends common.DatumLegend<D>
 
     // Show measures if [showMeasures] is true and there is a selection or if
     // showing measures when there is no selection.
-    final showMeasures = config.showMeasures &&
+    final showMeasures = config.showMeasures! &&
         (hasSelection ||
             legendDefaultMeasure != common.LegendDefaultMeasure.none);
 
-    return config.contentBuilder
+    return config.contentBuilder!
         .build(context, legendState, this, showMeasures: showMeasures);
   }
 
