@@ -30,10 +30,10 @@ class TabularLegendLayout implements LegendLayout {
   /// Default EdgeInsets for padding rows to the max column count
   static const defaultCellPadding = const EdgeInsets.all(8.0);
 
-  final bool isHorizontalFirst;
-  final int desiredMaxRows;
-  final int desiredMaxColumns;
-  final EdgeInsets cellPadding;
+  final bool? isHorizontalFirst;
+  final int? desiredMaxRows;
+  final int? desiredMaxColumns;
+  final EdgeInsets? cellPadding;
 
   TabularLegendLayout._internal(
       {this.isHorizontalFirst,
@@ -49,8 +49,8 @@ class TabularLegendLayout implements LegendLayout {
   ///
   /// [cellPadding] the [EdgeInsets] for each widget.
   factory TabularLegendLayout.horizontalFirst({
-    int desiredMaxColumns,
-    EdgeInsets cellPadding,
+    int? desiredMaxColumns,
+    EdgeInsets? cellPadding,
   }) {
     return new TabularLegendLayout._internal(
       isHorizontalFirst: true,
@@ -68,8 +68,8 @@ class TabularLegendLayout implements LegendLayout {
   ///
   /// [cellPadding] the [EdgeInsets] for each widget.
   factory TabularLegendLayout.verticalFirst({
-    int desiredMaxRows,
-    EdgeInsets cellPadding,
+    int? desiredMaxRows,
+    EdgeInsets? cellPadding,
   }) {
     return new TabularLegendLayout._internal(
       isHorizontalFirst: false,
@@ -84,10 +84,10 @@ class TabularLegendLayout implements LegendLayout {
     final paddedLegendEntries = ((cellPadding == null)
         ? legendEntries
         : legendEntries
-            .map((entry) => new Padding(padding: cellPadding, child: entry))
+            .map((entry) => new Padding(padding: cellPadding!, child: entry))
             .toList());
 
-    return isHorizontalFirst
+    return isHorizontalFirst!
         ? _buildHorizontalFirst(paddedLegendEntries)
         : _buildVerticalFirst(paddedLegendEntries);
   }
@@ -107,7 +107,7 @@ class TabularLegendLayout implements LegendLayout {
   Widget _buildHorizontalFirst(List<Widget> legendEntries) {
     final maxColumns = (desiredMaxColumns == _noLimit)
         ? legendEntries.length
-        : min(legendEntries.length, desiredMaxColumns);
+        : min(legendEntries.length, desiredMaxColumns!);
 
     final rows = <TableRow>[];
     for (var i = 0; i < legendEntries.length; i += maxColumns) {
@@ -123,12 +123,12 @@ class TabularLegendLayout implements LegendLayout {
   Widget _buildVerticalFirst(List<Widget> legendEntries) {
     final maxRows = (desiredMaxRows == _noLimit)
         ? legendEntries.length
-        : min(legendEntries.length, desiredMaxRows);
+        : min(legendEntries.length, desiredMaxRows!);
 
     final rows =
         new List.generate(maxRows, (_) => new TableRow(children: <Widget>[]));
     for (var i = 0; i < legendEntries.length; i++) {
-      rows[i % maxRows].children.add(legendEntries[i]);
+      rows[i % maxRows].children!.add(legendEntries[i]);
     }
 
     return _buildTableFromRows(rows);
@@ -140,11 +140,11 @@ class TabularLegendLayout implements LegendLayout {
     // Pad rows to the max column count, because each TableRow in a table is
     // required to have the same number of children.
     final columnCount = rows
-        .map((r) => r.children.length)
-        .fold(0, (max, current) => (current > max) ? current : max);
+        .map((r) => r.children!.length)
+        .fold(0, (dynamic max, current) => (current > max) ? current : max);
 
     for (var i = 0; i < rows.length; i++) {
-      final rowChildren = rows[i].children;
+      final rowChildren = rows[i].children!;
       final padCount = columnCount - rowChildren.length;
       if (padCount > 0) {
         rowChildren
