@@ -50,14 +50,14 @@ import 'chart_behavior.dart' show ChartBehavior, GestureType;
 /// removed.
 @immutable
 class SelectNearest extends ChartBehavior<common.SelectNearest> {
-  final Set<GestureType> desiredGestures;
+  final Set<GestureType>? desiredGestures;
 
-  final common.SelectionModelType selectionModelType;
-  final common.SelectionTrigger eventTrigger;
+  final common.SelectionModelType? selectionModelType;
+  final common.SelectionTrigger? eventTrigger;
   final common.SelectionMode selectionMode;
   final bool selectAcrossAllDrawAreaComponents;
   final bool selectClosestSeries;
-  final int maximumDomainDistancePx;
+  final int? maximumDomainDistancePx;
 
   SelectNearest._internal(
       {this.selectionModelType,
@@ -75,7 +75,7 @@ class SelectNearest extends ChartBehavior<common.SelectNearest> {
       bool selectAcrossAllDrawAreaComponents = false,
       bool selectClosestSeries = true,
       common.SelectionTrigger eventTrigger = common.SelectionTrigger.tap,
-      int maximumDomainDistancePx}) {
+      int? maximumDomainDistancePx}) {
     return new SelectNearest._internal(
         selectionModelType: selectionModelType,
         selectionMode: selectionMode,
@@ -114,11 +114,11 @@ class SelectNearest extends ChartBehavior<common.SelectNearest> {
   @override
   common.SelectNearest<D> createCommonBehavior<D>() {
     return new common.SelectNearest<D>(
-        selectionModelType: selectionModelType,
-        eventTrigger: eventTrigger,
+        selectionModelType: selectionModelType!,
+        eventTrigger: eventTrigger!,
         selectionMode: selectionMode,
         selectClosestSeries: selectClosestSeries,
-        maximumDomainDistancePx: maximumDomainDistancePx);
+        maximumDomainDistancePx: maximumDomainDistancePx!);
   }
 
   @override
