@@ -17,9 +17,9 @@ import 'package:charts_common/common.dart' as common show Color, LineStyle;
 
 class LineStyle implements common.LineStyle {
   @override
-  common.Color color;
+  late common.Color color;
   @override
-  List<int> dashPattern;
+  late List<int> dashPattern;
   @override
-  int strokeWidth;
+  late int strokeWidth;
 }
