@@ -32,9 +32,9 @@ class SymbolRendererCanvas implements SymbolRendererBuilder {
 
   @override
   Widget build(BuildContext context,
-      {Color color, Size size, bool enabled = true}) {
-    if (!enabled) {
-      color = color.withOpacity(0.26);
+      {Color? color, Size? size, bool? enabled = true}) {
+    if (!enabled!) {
+      color = color!.withOpacity(0.26);
     }
 
     return new SizedBox.fromSize(
@@ -55,15 +55,15 @@ abstract class CustomSymbolRenderer extends common.SymbolRenderer
   /// Must override this method to build the custom Widget with the given color
   /// as
   @override
-  Widget build(BuildContext context, {Color color, Size size, bool enabled});
+  Widget build(BuildContext context, {Color? color, Size? size, bool? enabled});
 
   @override
   void paint(common.ChartCanvas canvas, Rectangle<num> bounds,
-      {List<int> dashPattern,
-      common.Color fillColor,
-      common.FillPatternType fillPattern,
-      common.Color strokeColor,
-      double strokeWidthPx}) {
+      {List<int>? dashPattern,
+      common.Color? fillColor,
+      common.FillPatternType? fillPattern,
+      common.Color? strokeColor,
+      double? strokeWidthPx}) {
     // Intentionally ignored (never called).
   }
 
@@ -76,7 +76,7 @@ abstract class CustomSymbolRenderer extends common.SymbolRenderer
 /// Common interface for [CustomSymbolRenderer] & [SymbolRendererCanvas] for
 /// convenience for [LegendEntryLayout].
 abstract class SymbolRendererBuilder {
-  Widget build(BuildContext context, {Color color, Size size, bool enabled});
+  Widget build(BuildContext context, {Color? color, Size? size, bool? enabled});
 }
 
 /// The Widget which fulfills the guts of [SymbolRendererCanvas] actually
@@ -84,7 +84,7 @@ abstract class SymbolRendererBuilder {
 class _SymbolCustomPaint extends CustomPainter {
   final BuildContext context;
   final common.SymbolRenderer symbolRenderer;
-  final Color color;
+  final Color? color;
   final List<int> dashPattern;
 
   _SymbolCustomPaint(
@@ -95,7 +95,7 @@ class _SymbolCustomPaint extends CustomPainter {
     final bounds =
         new Rectangle<num>(0, 0, size.width.toInt(), size.height.toInt());
     final commonColor = new common.Color(
-        r: color.red, g: color.green, b: color.blue, a: color.alpha);
+        r: color!.red, g: color!.green, b: color!.blue, a: color!.alpha);
     symbolRenderer.paint(
         new ChartCanvas(canvas, GraphicsFactory(context)), bounds,
         fillColor: commonColor,
