@@ -57,7 +57,7 @@ class WidgetLayoutDelegate extends MultiChildLayoutDelegate {
             isRTL ? common.BehaviorPosition.end : common.BehaviorPosition.start;
         final rightPosition =
             isRTL ? common.BehaviorPosition.start : common.BehaviorPosition.end;
-        final behaviorPosition = idAndBehavior[behaviorID].position;
+        final behaviorPosition = idAndBehavior[behaviorID]!.position;
 
         behaviorSize = layoutChild(behaviorID, new BoxConstraints.loose(size));
         if (behaviorPosition == common.BehaviorPosition.top) {
@@ -86,8 +86,8 @@ class WidgetLayoutDelegate extends MultiChildLayoutDelegate {
       // TODO: Unable to relayout with new smaller width.
       // In the delegate, all children are required to have layout called
       // exactly once.
-      final behaviorOffset = _getBehaviorOffset(idAndBehavior[behaviorID],
-          behaviorSize: behaviorSize, chartSize: chartSize, isRTL: isRTL);
+      final behaviorOffset = _getBehaviorOffset(idAndBehavior[behaviorID]!,
+          behaviorSize: behaviorSize, chartSize: chartSize, isRTL: isRTL)!;
 
       positionChild(behaviorID, behaviorOffset);
     }
@@ -101,9 +101,9 @@ class WidgetLayoutDelegate extends MultiChildLayoutDelegate {
   }
 
   // Calculate buildable behavior's offset.
-  Offset _getBehaviorOffset(BuildableBehavior behavior,
-      {Size behaviorSize, Size chartSize, bool isRTL}) {
-    Offset behaviorOffset;
+  Offset? _getBehaviorOffset(BuildableBehavior behavior,
+      {Size? behaviorSize, Size? chartSize, bool? isRTL}) {
+    Offset? behaviorOffset;
 
     final behaviorPosition = behavior.position;
     final outsideJustification = behavior.outsideJustification;
@@ -112,7 +112,7 @@ class WidgetLayoutDelegate extends MultiChildLayoutDelegate {
     if (behaviorPosition == common.BehaviorPosition.top ||
         behaviorPosition == common.BehaviorPosition.bottom) {
       final heightOffset = behaviorPosition == common.BehaviorPosition.bottom
-          ? chartSize.height
+          ? chartSize!.height
           : 0.0;
 
       final horizontalJustification =
@@ -128,19 +128,19 @@ class WidgetLayoutDelegate extends MultiChildLayoutDelegate {
           break;
         case _HorizontalJustification.rightDrawArea:
           behaviorOffset = new Offset(
-              behavior.drawAreaBounds.right - behaviorSize.width, heightOffset);
+              behavior.drawAreaBounds.right - behaviorSize!.width, heightOffset);
           break;
         case _HorizontalJustification.right:
           behaviorOffset =
-              new Offset(chartSize.width - behaviorSize.width, heightOffset);
+              new Offset(chartSize!.width - behaviorSize!.width, heightOffset);
           break;
       }
     } else if (behaviorPosition == common.BehaviorPosition.start ||
         behaviorPosition == common.BehaviorPosition.end) {
       final widthOffset =
-          (isRTL && behaviorPosition == common.BehaviorPosition.start) ||
+          (isRTL! && behaviorPosition == common.BehaviorPosition.start) ||
                   (!isRTL && behaviorPosition == common.BehaviorPosition.end)
-              ? chartSize.width
+              ? chartSize!.width
               : 0.0;
 
       switch (outsideJustification) {
@@ -155,22 +155,22 @@ class WidgetLayoutDelegate extends MultiChildLayoutDelegate {
           break;
         case common.OutsideJustification.endDrawArea:
           behaviorOffset = new Offset(widthOffset,
-              behavior.drawAreaBounds.bottom - behaviorSize.height);
+              behavior.drawAreaBounds.bottom - behaviorSize!.height);
           break;
         case common.OutsideJustification.end:
           behaviorOffset =
-              new Offset(widthOffset, chartSize.height - behaviorSize.height);
+              new Offset(widthOffset, chartSize!.height - behaviorSize!.height);
           break;
       }
     } else if (behaviorPosition == common.BehaviorPosition.inside) {
-      var rightOffset = new Offset(chartSize.width - behaviorSize.width, 0.0);
+      var rightOffset = new Offset(chartSize!.width - behaviorSize!.width, 0.0);
 
       switch (insideJustification) {
         case common.InsideJustification.topStart:
-          behaviorOffset = isRTL ? rightOffset : Offset.zero;
+          behaviorOffset = isRTL! ? rightOffset : Offset.zero;
           break;
         case common.InsideJustification.topEnd:
-          behaviorOffset = isRTL ? Offset.zero : rightOffset;
+          behaviorOffset = isRTL! ? Offset.zero : rightOffset;
           break;
       }
     }
@@ -178,30 +178,30 @@ class WidgetLayoutDelegate extends MultiChildLayoutDelegate {
     return behaviorOffset;
   }
 
-  _HorizontalJustification getOutsideJustification(
-      common.OutsideJustification justification, bool isRTL) {
-    _HorizontalJustification mappedJustification;
+  _HorizontalJustification? getOutsideJustification(
+      common.OutsideJustification justification, bool? isRTL) {
+    _HorizontalJustification? mappedJustification;
 
     switch (justification) {
       case common.OutsideJustification.startDrawArea:
       case common.OutsideJustification.middleDrawArea:
-        mappedJustification = isRTL
+        mappedJustification = isRTL!
             ? _HorizontalJustification.rightDrawArea
             : _HorizontalJustification.leftDrawArea;
         break;
       case common.OutsideJustification.start:
       case common.OutsideJustification.middle:
-        mappedJustification = isRTL
+        mappedJustification = isRTL!
             ? _HorizontalJustification.right
             : _HorizontalJustification.left;
         break;
       case common.OutsideJustification.endDrawArea:
-        mappedJustification = isRTL
+        mappedJustification = isRTL!
             ? _HorizontalJustification.leftDrawArea
             : _HorizontalJustification.rightDrawArea;
         break;
       case common.OutsideJustification.end:
-        mappedJustification = isRTL
+        mappedJustification = isRTL!
             ? _HorizontalJustification.left
             : _HorizontalJustification.right;
         break;
