@@ -22,7 +22,7 @@ import 'package:charts_flutter/src/behaviors/legend/legend_layout.dart';
 class MockContext extends Mock implements BuildContext {}
 
 void main() {
-  BuildContext context;
+  late BuildContext context;
 
   setUp(() {
     context = new MockContext();
@@ -33,9 +33,9 @@ void main() {
       final builder = new TabularLegendLayout.horizontalFirst();
       final widgets = <Widget>[new Text('1'), new Text('2'), new Text('3')];
 
-      final Table layout = builder.build(context, widgets);
+      final Table layout = builder.build(context, widgets) as Table;
       expect(layout.children.length, 1);
-      expect(layout.children.first.children.length, 3);
+      expect(layout.children.first.children!.length, 3);
     });
 
     test('does not build extra columns if max columns exceed widget count', () {
@@ -43,9 +43,9 @@ void main() {
           new TabularLegendLayout.horizontalFirst(desiredMaxColumns: 10);
       final widgets = <Widget>[new Text('1'), new Text('2'), new Text('3')];
 
-      final Table layout = builder.build(context, widgets);
+      final Table layout = builder.build(context, widgets) as Table;
       expect(layout.children.length, 1);
-      expect(layout.children.first.children.length, 3);
+      expect(layout.children.first.children!.length, 3);
     });
 
     test('builds horizontally until max column exceeded', () {
@@ -55,41 +55,41 @@ void main() {
       final widgets = new List<Widget>.generate(
           7, (int index) => new Text(index.toString()));
 
-      final Table layout = builder.build(context, widgets);
+      final Table layout = builder.build(context, widgets) as Table;
       expect(layout.children.length, 4);
 
-      expect(layout.children[0].children[0], equals(widgets[0]));
-      expect(layout.children[0].children[1], equals(widgets[1]));
+      expect(layout.children[0].children![0], equals(widgets[0]));
+      expect(layout.children[0].children![1], equals(widgets[1]));
 
-      expect(layout.children[1].children[0], equals(widgets[2]));
-      expect(layout.children[1].children[1], equals(widgets[3]));
+      expect(layout.children[1].children![0], equals(widgets[2]));
+      expect(layout.children[1].children![1], equals(widgets[3]));
 
-      expect(layout.children[2].children[0], equals(widgets[4]));
-      expect(layout.children[2].children[1], equals(widgets[5]));
+      expect(layout.children[2].children![0], equals(widgets[4]));
+      expect(layout.children[2].children![1], equals(widgets[5]));
 
-      expect(layout.children[3].children[0], equals(widgets[6]));
+      expect(layout.children[3].children![0], equals(widgets[6]));
     });
 
     test('builds vertically', () {
       final builder = new TabularLegendLayout.verticalFirst();
       final widgets = <Widget>[new Text('1'), new Text('2'), new Text('3')];
 
-      final Table layout = builder.build(context, widgets);
+      final Table layout = builder.build(context, widgets) as Table;
       expect(layout.children.length, 3);
-      expect(layout.children[0].children.length, 1);
-      expect(layout.children[1].children.length, 1);
-      expect(layout.children[2].children.length, 1);
+      expect(layout.children[0].children!.length, 1);
+      expect(layout.children[1].children!.length, 1);
+      expect(layout.children[2].children!.length, 1);
     });
 
     test('does not build extra rows if max rows exceed widget count', () {
       final builder = new TabularLegendLayout.verticalFirst(desiredMaxRows: 10);
       final widgets = <Widget>[new Text('1'), new Text('2'), new Text('3')];
 
-      final Table layout = builder.build(context, widgets);
+      final Table layout = builder.build(context, widgets) as Table;
       expect(layout.children.length, 3);
-      expect(layout.children[0].children.length, 1);
-      expect(layout.children[1].children.length, 1);
-      expect(layout.children[2].children.length, 1);
+      expect(layout.children[0].children!.length, 1);
+      expect(layout.children[1].children!.length, 1);
+      expect(layout.children[2].children!.length, 1);
     });
 
     test('builds vertically until max column exceeded', () {
@@ -98,19 +98,19 @@ void main() {
       final widgets = new List<Widget>.generate(
           7, (int index) => new Text(index.toString()));
 
-      final Table layout = builder.build(context, widgets);
+      final Table layout = builder.build(context, widgets) as Table;
       expect(layout.children.length, 2);
 
-      expect(layout.children[0].children[0], equals(widgets[0]));
-      expect(layout.children[1].children[0], equals(widgets[1]));
+      expect(layout.children[0].children![0], equals(widgets[0]));
+      expect(layout.children[1].children![0], equals(widgets[1]));
 
-      expect(layout.children[0].children[1], equals(widgets[2]));
-      expect(layout.children[1].children[1], equals(widgets[3]));
+      expect(layout.children[0].children![1], equals(widgets[2]));
+      expect(layout.children[1].children![1], equals(widgets[3]));
 
-      expect(layout.children[0].children[2], equals(widgets[4]));
-      expect(layout.children[1].children[2], equals(widgets[5]));
+      expect(layout.children[0].children![2], equals(widgets[4]));
+      expect(layout.children[1].children![2], equals(widgets[5]));
 
-      expect(layout.children[0].children[3], equals(widgets[6]));
+      expect(layout.children[0].children![3], equals(widgets[6]));
     });
   });
 }
