@@ -23,16 +23,16 @@ import 'selection_model_config.dart' show SelectionModelConfig;
 class PieChart<D> extends BaseChart<D> {
   PieChart(
     List<common.Series> seriesList, {
-    bool animate,
-    Duration animationDuration,
-    common.ArcRendererConfig<D> defaultRenderer,
-    List<ChartBehavior> behaviors,
-    List<SelectionModelConfig<D>> selectionModels,
-    common.RTLSpec rtlSpec,
-    LayoutConfig layoutConfig,
+    bool? animate,
+    Duration? animationDuration,
+    common.ArcRendererConfig<D>? defaultRenderer,
+    List<ChartBehavior>? behaviors,
+    List<SelectionModelConfig<D>>? selectionModels,
+    common.RTLSpec? rtlSpec,
+    LayoutConfig? layoutConfig,
     bool defaultInteractions = true,
   }) : super(
-          seriesList,
+          seriesList as List<Series<dynamic, D*>>,
           animate: animate,
           animationDuration: animationDuration,
           defaultRenderer: defaultRenderer,
@@ -44,8 +44,8 @@ class PieChart<D> extends BaseChart<D> {
         );
 
   @override
-  common.PieChart<D> createCommonChart(BaseChartState chartState) =>
-      new common.PieChart<D>(layoutConfig: layoutConfig?.commonLayoutConfig);
+  common.PieChart<D> createCommonChart(BaseChartState? chartState) =>
+      new common.PieChart<D>((layoutConfig: layoutConfig?.commonLayoutConfig)!);
 
   @override
   void addDefaultInteractions(List<ChartBehavior> behaviors) {
