@@ -27,7 +27,7 @@ void main() {
           new charts.SeriesDatumConfig<String>('Sales', '2016')
         ]);
 
-    charts.SelectionModel<String> currentSelectionModel;
+    charts.SelectionModel<String>? currentSelectionModel;
 
     void selectionChangedListener(charts.SelectionModel<String> model) {
       currentSelectionModel = model;
@@ -48,13 +48,13 @@ void main() {
 
     await tester.pump();
 
-    expect(currentSelectionModel.selectedDatum, hasLength(1));
+    expect(currentSelectionModel!.selectedDatum, hasLength(1));
     final selectedDatum =
-        currentSelectionModel.selectedDatum.first.datum as OrdinalSales;
+        currentSelectionModel!.selectedDatum.first.datum as OrdinalSales;
     expect(selectedDatum.year, equals('2016'));
     expect(selectedDatum.sales, equals(100));
-    expect(currentSelectionModel.selectedSeries, hasLength(1));
-    expect(currentSelectionModel.selectedSeries.first.id, equals('Sales'));
+    expect(currentSelectionModel!.selectedSeries, hasLength(1));
+    expect(currentSelectionModel!.selectedSeries.first.id, equals('Sales'));
   });
 }
 
